Split comparison.scala in smaller files
